chore(dep-resolution): add error logging for dynamic resolution (semgrep/semgrep-proprietary#2440)
Adds new support for errors that occur during the updated dependency
resolution process. The errors that are currently supported are all
WARNING level, so will appear only when using `--verbose`.

Note that the error messages produced contain the entirety of the output
from the call to the package manager, so they can get quite long.

As part of these changes, we add a new utility to `UCmd.ml` that
captures stderr as well as stdout for a command run. This allows us to
include the stderr output in the error that we generate and therefore in
the log that we show to the user.

# Test plan
For now, manual testing only. The method for manual testing was:
- Create a local gradle project and specify the Java version to be
something that is not present on my machine
- Run a scan with `semgrep ci --allow-dynamic-dependency-resolution
--verbose`
- Verify that the error from gradle is produced in the CLI output

I attempted to create an automated test for the python-side handling of
the returned errors via a snapshot e2e test, but ran into difficulty
mocking the RPC call. I moved the test into a separate PR at
https://github.com/semgrep/semgrep-proprietary/pull/2534.

synced from Pro 1bdd3089d49b2694585cbe021bb01595e1aebf0f