remove `PYINSTALLER_RESET_ENVIRONMENT = 1`
Since we now use subprocesses on Windows, and we wait for completion, we
do not need this any more. It was added because of `execvp` behaviour
on Windows where the process would exit after spawning another with
`execvp`, and the _MEIXXXX folder was no longer available, leading to
strange runtime errors.