protect `print_cli_progress` with mutex
This is needed when using Domains, else the output gets messed up and we
get random errors. Specifically, flushing seems to happen in an
interleaving way, so we get lines with 2 dots and empty newlines, which
confuses the progress bar.

This should be refactor as it's very fragile, but for now this fixes it.