fix: timeout threads were delaying the result
We need to make sure that the threads implementing timeouts are stopped
when the actual work is done, else they linger and delay the whole
program.

This is achieved by breaking down the sleeping duration in smaller
periods of some small granularity, and checking an atomic variable to
see if we need to exit or sleep more until the max duration is reached,
which will trigger cancellation of the work task.