fix: ensure that shared references are in DLS
The function `Common.save_excursion` is used in a lot of places to
temporarily change the value of a local or global reference (or flag).
This leads to a large number of dangerous races.

For now the solution is as follows:

- Add an extra function `Common.save_excursion_unsafe` which can be used
when the value affected is local, or is a global that we consider to be
used in a safe way even in the presence of this function.
- Make the original function `Common.save_excursion` take a
`Domain.DLS.key` instaed of a `ref`. All globals and flags affected have
been converted.

This does not solve all thread-safety issues!