optimise: cache defs once calculated
The defs come from a file set in a flag and nowhere else, and are common
to all C files parsed in a scan.

Therefore, there is no need calculate them more than once.

The current solution allows for some redundant work to take place in
case of races to create the table and set the cache, but it remains
safe since the values are always the same.