add comments on some `Log.warn` invocations
The TL;DR is that it's probablçy not worth it to push some calculations
that are only needed for the log message into the logging function. The
reason is that warnings are emitted by default and the logging function
is protected with a mutex, so it could have a detrimental effect on
performance during normal operation.

Most of these operations seem pretty cheap anyway.