optimise allocations: prefer `Hashtbl.find`
To my suprise, doing something like:

```
match Hashtbl.find tbl key with
| v -> ...
| exception Not_found -> ...
```

is not really slower, in most cases, than:

```
match Hashtbl.find_opt tbl key with
| Some v -> ...
| None -> ...
```

and it allocated a LOT less.

Initially I expected the exception to be very expensive, but I was
wrong. Tried in utop but also with a natively compiled benchmark.

Of course there's a dependency to the distribution of keys, how many are
expected to be missing etc, but in my experiments I did not observe
drastic changes even when most keys were not found and the exception was
raised.

So in this commit I changed those instances where I suspect that
allocation can be heavy, without doing a proper benchmark.