Warn if --output-enclosing-context without --json
At the moment, the enclosing context captured when run with --output-enclosing-context show up only in the --json output. So, if the user specifies the former but not the latter, we emit a warning that the former will have no effect.

NB. Adding enclosing context in other outputs should be easy and can be added to opengrep on-demand