Show enclosing context when focusing on a metavar (partial solution)
The solution is partial, because when we focus on a metavariable, but the rule captures the entire function, we do not include this function in the context, for example:

rule:
```
- pattern: foo() { ... $VAR = 5; ...}
- focus-metavariable: $VAR
```

code:
```
class C {
  foo() { x = 5; }
}
```

In the above, the rule matches the entire definition of foo(), but we focus on x only. In such a case, we want the context to be `class C` and `function foo`, but we only see `class C` in teh enclosing context. This is because currently the context is for the match of the rule.

At the moment, the enclosing context is constructed in the visitor, so for each match. If the match captures, say, a function, this function is no added to the enclosing context, so when we focus on a metavariable in the example above, we strip the result from `foo() { ... = 5; }` and only report `x` at a given range with the enclosing context only `class C`. To fix this, we would need to change the algorithm of matching to keep the enclosing context at all times or recalculate it for each match. TBD.