add `force-exclude` flag: apply exclude to files
The default behaviour is to not apply exclude and include filtering to
targets passed to the scan (or ci) command if they are files. The
filtering only takes place for files under targets that are
directories.

This flag makes the behaviour more uniform, applying the filters to all
targets.

See #264.