windows: fix #291 cwd can begin with lowercase
This happens when opengrep is invoked as subprocess in integrations, and
in some of these cases `Unix.getcwd` and `Sys.getcwd` can return the
path with lowercase drive letter.

This is easy to check in utop: execute `Unix.chdir "c:\\..."` with some
path then `Unix.getcwd ()` will return the path starting with a
lowercase drive. (Same for `Sys.getcwd ()`).

I hesitate to normalise the paths because apparently, in some Windows
filesystems like NTFS it's possible to have case-sensitive paths.

Therefore we make do with this small hack, which only has effect on
Windows and only normalises the drive letter.